reuse page_config and submenus
